update Category API

+++ app/Http/Controllers/Api/Category/CategoryV1Controller.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\Api\Category;
 use App\Http\Controllers\Api\Category\CategoryTransformer;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use Dingo\Api\Exception\StoreResourceFailedException;
 use Dingo\Api\Routing\Helpers;
-use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 
@@ -35,30 +35,27 @@ class CategoryV1Controller extends Controller
 
     public function store(Request $request)
     {
-        try {
-            $category = Category::create([
-                'name' => $request->name,
-                'slug' => Str::slug($request->name),
-                'description' => $request->description,
-                'image' => $request->image,
-                'parent_id' => $request->parent_id,
-                'status' => $request->status,
-                'keywords' => $request->keywords,
-            ]);
+        $rules = [
+            'name' => ['required', 'string'],
+            'image' => ['required', 'string'],
+            'status' => ['required', 'integer'],
+            'slug' => ['unique:categories'],
+            'parent_id' => ['nullable', 'integer'],
+            'keywords' => ['nullable', 'string'],
+            'description' => ['nullable', 'string'],
+        ];
 
-            return $this->response->item($category, new CategoryTransformer());
-        } catch (Exception $ex) {
-            return $this->response->error($ex->getMessage(), 500);
+        $payload = app('request')->only(['name', 'slug', 'status', 'description', 'image', 'parent_id', 'keywords']);
+
+        $validator = app('validator')->make($payload, $rules);
+
+        if ($validator->fails()) {
+            throw new StoreResourceFailedException('Could not create new Category.', $validator->errors());
         }
-    }
 
-    public function update(Request $request, $id)
-    {
-        $category = Category::find($id);
-
-        $category->update([
+        $category = Category::create([
             'name' => $request->name,
-            'slug' => $request->slug,
+            'slug' => $request->slug ? $request->slug : Str::slug($request->name),
             'description' => $request->description,
             'image' => $request->image,
             'parent_id' => $request->parent_id,
@@ -66,15 +63,64 @@ class CategoryV1Controller extends Controller
             'keywords' => $request->keywords,
         ]);
 
-        return $this->response->item($category, new CategoryTransformer());
+        return $this->response->array([
+            'message' => 'Category created successfully',
+        ]);
+    }
+
+    public function update(Request $request, $id)
+    {
+        $rules = [
+            'name' => ['required', 'string'],
+            'image' => ['string'],
+            'status' => ['required', 'integer'],
+            'slug' => ['unique:categories'],
+            'parent_id' => ['nullable', 'integer'],
+            'keywords' => ['nullable', 'string'],
+            'description' => ['nullable', 'string'],
+        ];
+
+        $payload = app('request')->only(['name', 'slug', 'status', 'description', 'image', 'parent_id', 'keywords']);
+
+        $validator = app('validator')->make($payload, $rules);
+
+        if ($validator->fails()) {
+            throw new StoreResourceFailedException('Could not update Category.', $validator->errors());
+        }
+
+        $category = Category::find($id);
+
+        if (!$category) {
+            return $this->response->errorNotFound('Category not found');
+        }
+
+        $category->update([
+            'name' => $request->name,
+            'slug' => $request->slug ? $request->slug : Str::slug($request->name),
+            'description' => $request->description,
+            'image' => $request->image,
+            'parent_id' => $request->parent_id,
+            'status' => $request->status,
+            'keywords' => $request->keywords,
+        ]);
+
+        return $this->response->array([
+            'message' => 'Category updated successfully',
+        ]);
     }
 
     public function destroy($id)
     {
         $category = Category::find($id);
 
+        if (!$category) {
+            return $this->response->errorNotFound('Category not found');
+        }
+
         $category->delete();
 
-        return $this->response->noContent();
+        return $this->response->array([
+            'message' => 'Category deleted successfully',
+        ]);
     }
 }
